refactor: split reminder ui facade
